refactor-优化ID获取方式

+++ src/main/java/com/example/demo/player/dto/AddPlayerExperienceReq.java
@@ -20,6 +20,6 @@ public class AddPlayerExperienceReq {
 	/**
 	 * 新增经验值
 	 */
-	private int addValue;
+	private long addValue;
 
 }
